(Numerous conflicts manually updated -- trunk further diverges due to new features in pkgng)
Update PostgreSQL ports to latest version.

Data Corruption Fix

For users of PostgreSQL versions 9.3 or 9.4, this release fixes a problem where
the database will fail to protect against "multixact wraparound", resulting in
data corruption or loss. Users with a high transaction rate (1 million or more
per hour) in a database with many foreign keys are especially vulnerable. We
strongly urge all users of 9.4 and 9.3 to update their installations in the
next few days.

Users of versions 9.2 and earlier are not affected by this issue.

Security:	fc38cd83-00b3-11e5-8ebd-0026551a22dc

Conflicts:
	databases/postgresql90-client/Makefile
	databases/postgresql90-server/Makefile
	databases/postgresql91-client/Makefile
	databases/postgresql91-server/pkg-plist-contrib
	databases/postgresql92-client/Makefile
	databases/postgresql92-server/pkg-plist-contrib
	databases/postgresql93-client/Makefile
	databases/postgresql93-server/pkg-plist-contrib
	databases/postgresql94-client/Makefile